// src/api/types.gen.ts
export type UserLoginType = 'TRADITIONAL' | 'GOOGLE';

export type UserRole = 'ADMIN' | 'CLIENT' | 'CONTRACTOR' | 'AGENT' | 'CONSULTANCY';

export type MediaDocumentType = 'CITIZENSHIP_FRONT' | 'CITIZENSHIP_BACK' | 'PASSPORT' | 'PROFILE_IMAGE' | 'COMPANY_REGISTRATION' | 'TAX_REGISTRATION' | 'COMPANY_LOGO' | 'PERSONAL_LICENSE' | 'INSURANCE_DOCUMENT' | 'PROJECT_DOCUMENT' | 'PORTFOLIO_PROJECT_PHOTO' | 'BID_DOCUMENT';

export type MediaEntityType = 'PROJECT' | 'CLIENT_PROFILE' | 'CONTRACTOR_PROFILE' | 'CONSULTANCY_PROFILE' | 'CONTRACTOR_PORTFOLIO_PROJECT' | 'CONSULTANCY_PORTFOLIO_PROJECT' | 'BID';

export type User = {
    id: string;
    createdAt: string;
    updatedAt: string;
    deletedAt: string;
    fullName: string;
    email: string;
    location?: string;
    password: string;
    phone?: string;
    isEmailVerified?: boolean;
    loginType?: UserLoginType;
    address?: string;
    avatar?: string;
    role: UserRole;
    profileImage?: Media;
    tokens?: Array<Token>;
    logs: Array<Log>;
    projects: Array<Project>;
    accountVerifications: Array<AccountVerification>;
    clientProfile: ClientProfile;
    contractorProfile: ContractorProfile;
    designConsultancyProfile: DesignConsultancyProfile;
    agentProfile?: AgentProfile;
};

export type Media = {
    id: string;
    createdAt: string;
    updatedAt: string;
    deletedAt: string;
    mimeType: string;
    name: string;
    documentType: MediaDocumentType;
    entityType: MediaEntityType;
    entityId?: string;
    originalFileName: string;
    url: string;
    user: User;
};

export type TokenEnum = 'ACCESS' | 'REFRESH' | 'FORGET_PASSWORD' | 'VERIFY_EMAIL' | 'CHANGE_PASSWORD_OTP';

export type Token = {
    id: string;
    createdAt: string;
    updatedAt: string;
    deletedAt: string;
    tokenHash: string;
    type: TokenEnum;
    deviceId?: string;
    ipHash?: string;
    expiresAt: string;
    usedAt?: string;
    revokedAt?: string;
    user?: User;
};
// ...
export type Log = {
    id: string;
    createdAt: string;
    updatedAt: string;
    deletedAt: string;
    module: LogModule;
    action: LogAction;
    severity: LogSeverity;
    actorType: ActorType;
    details: RecordStringAny;
    ipAddress: string;
    userAgent: string;
    resource: string;
    description: string;
    errorMessage: string;
    errorStack: string;
    user: User;
};

export type Project = {
    id: string;
    createdAt: string;
    updatedAt: string;
    deletedAt: string;
    createdBy: User;
    awardedContractor?: User;
    bids: Array<Bids>;
    projectDocuments: Array<Media>;
    title: string;
    slug: string;
    description: string;
    category: ProjectCategory;
    location: string;
    budgetType?: ProjectBudgetType;
    minimumBudget?: number;
    maximumBudget?: number;
    target: ProjectTarget;
    preferredStartDate: string;
    preferredEndDate: string;
    bidStartedDate?: string;
    bidEndedDate?: string;
    experienceLevel?: ContractorExperienceLevel;
    similarProjectExperience?: string;
    additionalRequirements?: string;
    status: ProjectStatus;
    cancellationReason?: string;
};

export type ContractorProfile = {
    id: string;
    createdAt: string;
    updatedAt: string;
    deletedAt: string;
    user: User;
    companyName: string;
    yearsOfExperience?: number;
    specializations: string;
    teamSize: string;
    isAvailable: boolean;
    companyDescription: string;
    serviceAreas: Array<string>;
    companyRegistrationCertificate: Array<Media>;
    taxRegistrationCertificate: Array<Media>;
    personalLicense: Array<Media>;
    companyLogo: Array<Media>;
    insuranceDocument: Array<Media>;
    portfolioProjects: Array<PortfolioProject>;
};

export type PortfolioProject = {
    id: string;
    createdAt: string;
    updatedAt: string;
    deletedAt: string;
    contractor?: ContractorProfile;
    designConsultancy?: DesignConsultancyProfile;
    projectPhotos: Array<Media>;
    projectName: string;
    description: string;
    notableClients?: string;
    averageProjectValue?: number;
};

export type DesignConsultancyProfile = {
    id: string;
    createdAt: string;
    updatedAt: string;
    deletedAt: string;
    user: User;
    companyName: string;
    yearsOfExperience?: number;
    specializations: string;
    teamSize: number;
    companyDescription: string;
    serviceAreas: Array<string>;
    address?: string;
    isAvailable: boolean;
    companyRegistrationCertificate: Array<Media>;
    taxRegistrationCertificate: Array<Media>;
    companyLogo: Array<Media>;
    portfolioProjects: Array<PortfolioProject>;
};
// ...
export type AccountVerification = {
    id: string;
    createdAt: string;
    updatedAt: string;
    deletedAt: string;
    user: User;
    actionBy?: User;
    status: UserVerificationStatus;
    comments?: string;
    verifiedAt?: string;
};

export type ClientType = 'INDIVIDUAL' | 'COMPANY';

export type ClientProfile = {
    id: string;
    createdAt: string;
    updatedAt: string;
    deletedAt: string;
    user: User;
    clientType: ClientType;
    /**
     * Only for company clients
     */
    companyName?: string;
    billingAddress?: string;
    contactPerson?: string;
    vatNumber?: string;
    companyRegistrationNo?: string;
    panNumber?: string;
    extraData?: unknown;
    citizenshipCertificate: Array<Media>;
};

export type AgentProfile = {
    id: string;
    createdAt: string;
    updatedAt: string;
    deletedAt: string;
    user: User;
    hireDate?: string;
};
// ...
export type Blog = {
    id: string;
    createdAt: string;
    updatedAt: string;
    deletedAt: string;
    title: string;
    slug: string;
    content: string;
    isPublished: boolean;
    author: User;
    authorId: string;
    featuredImage?: Media;
    featuredImageId?: string;
    category?: string;
    tags?: Array<string>;
};